Refactor: Change structure of assets
* Create cards directory

* Create index.ts in each subdirectory

* Rename card-bg to default-bg

* Remove index.ts in assets directory
