+++ src/assets/cards/index.ts
@@ -1,0 +1,5 @@
+import DEFAULT_BG from './default-bg.jpg';
+
+export {
+    DEFAULT_BG
+}
