+++ src/assets/icons/index.ts
@@ -1,0 +1,26 @@
+import BottomChevron from './bottom-chevron.svg';
+import LeftChevron from './left-chevron.svg';
+import RightChevron from './right-chevron.svg';
+
+import CloseIcon from './close-icon.svg'
+import DownloadIcon from './download-icon.svg';
+import EmptyHeartIcon from './empty-heart-icon.svg';
+import FillHeartIcon from './fill-heart-icon.svg';
+import LoginIcon from './login-icon.svg';
+import LogoutIcon from './logout-icon.svg';
+import UserIcon from './user-icon.svg';
+import GithubIcon from './github-logo.svg';
+
+export {
+    BottomChevron,
+    LeftChevron,
+    RightChevron,
+    CloseIcon,
+    DownloadIcon,
+    EmptyHeartIcon,
+    FillHeartIcon,
+    LoginIcon,
+    LogoutIcon,
+    UserIcon,
+    GithubIcon
+}
