--- src/assets/index.ts
@@ -1,39 +1,0 @@
-// ICONS, 컴포넌트 형태
-import BottomChevron from './icons/bottom-chevron.svg';
-import LeftChevron from './icons/left-chevron.svg';
-import RightChevron from './icons/right-chevron.svg';
-
-import CloseIcon from './icons/close-icon.svg'
-import DownloadIcon from './icons/download-icon.svg';
-import EmptyHeartIcon from './icons/empty-heart-icon.svg';
-import FillHeartIcon from './icons/fill-heart-icon.svg';
-import LoginIcon from './icons/login-icon.svg';
-import LogoutIcon from './icons/logout-icon.svg';
-import UserIcon from './icons/user-icon.svg';
-import GithubIcon from './icons/github-logo.svg';
-
-// LOGO
-import FACEBOOK_LOGO from './logos/facebook-logo.png';
-import GITHUB_LOGO from './logos/github-logo.png';
-import GOOGLE_LOGO from './logos/google-logo.png';
-
-// ETC
-import CARD_BG from './card-bg.jpg';
-
-export {
-    BottomChevron,
-    LeftChevron,
-    RightChevron,
-    CloseIcon,
-    DownloadIcon,
-    EmptyHeartIcon,
-    FillHeartIcon,
-    LoginIcon,
-    LogoutIcon,
-    UserIcon,
-    GithubIcon,
-    FACEBOOK_LOGO,
-    GITHUB_LOGO,
-    GOOGLE_LOGO,
-    CARD_BG
-}
