+++ src/assets/logos/index.ts
@@ -1,0 +1,9 @@
+import FACEBOOK_LOGO from './facebook-logo.png';
+import GITHUB_LOGO from './github-logo.png';
+import GOOGLE_LOGO from './google-logo.png';
+
+export {
+    FACEBOOK_LOGO,
+    GITHUB_LOGO,
+    GOOGLE_LOGO
+}
